% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkGuess.R
\name{checkGuess}
\alias{checkGuess}
\title{Check guess}
\usage{
checkGuess(guess)
}
\arguments{
\item{guess}{A vector of character values containing the guessed colours.}
}
\value{
A list with the correctly guessed colours and slots.
}
\description{
Checks if a guess is correct. Is an internal function and not supposed to be called by the user.
}
