% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playGame.R
\name{playGame}
\alias{playGame}
\title{Play console game}
\usage{
playGame(seed, replaceColours = FALSE)
}
\arguments{
\item{seed}{Seed used when generating random numbers. Primarily used for testing purposes.}

\item{replaceColours}{Logical value describing if the same colour is allowed more than once. So far, only FALSE is implemented.}
}
\value{
Nothing, launches a console game
}
\description{
Launches a game that you can play on the console. Mainly used for testing.
}
