% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playShinyGame.R
\name{playShinyGame}
\alias{playShinyGame}
\title{Play Shiny game}
\usage{
playShinyGame(replaceColours = FALSE)
}
\arguments{
\item{replaceColours}{Logical value describing if the same colour is allowed more than once. So far, only FALSE is implemented.}
}
\value{
Nothing, launches the shiny app
}
\description{
Launches a game that you can play as a Shiny app.
}
